Solved crash on game object destroy

+++ src/Physics/PhysicsSystem.cpp
@@ -101,8 +101,8 @@ MeshCollider* PhysicsSystem::CreateHeightFieldMap(float* heightField, unsigned i
     
     unsigned int mapSize = width * height;
     
-    int maximumHeight =  1000;
-    int minimumHeight = -1000;
+    float maximumHeight =  1000;
+    float minimumHeight = -1000;
     
     collider->heightMapBuffer = new float[ mapSize ];
